pyproject.toml
[project]
name = "openjuris-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.129.0",
    "ftfy>=6.3.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "pydantic-settings>=2.13.0",
    "python-dateutil>=2.9.0.post0",
    "ruff>=0.15.1",
    "sentence-transformers>=5.2.3",
    "sqlalchemy-libsql>=0.2.0",
    "sqlmodel>=0.0.34",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
